refactor(cpp26): type-safe enums, lazy spdlog formatting, constexpr names
- Change MainOptions::algorithm from optional<int> to optional<LPAlgo>,
  eliminating all unsafe static_cast<LPAlgo>(int) throughout the CLI,
  config file, and --set overlay pipeline
- Range-check LPAlgo cast in gtopt_json_io_set.cpp against last_algo
  sentinel; reject invalid SolverLogMode with error instead of UB
- Replace 36 spdlog::info(std::format(...)) with native spdlog::info()
  across gtopt_lp_runner, lp_stats, json_io_parse, json_io_write, and
  standalone/main to avoid eager allocation when log level is disabled
- Replace SPDLOG_DEBUG(fmt::format(...)) in generator_lp.cpp
- Replace 12 magic string literals in flow_right_lp.cpp with constexpr
  constants (FlowName, FailName, QehName, DemandName, QavgName)
- Convert iterator loops in benders_cut.cpp to structured bindings
- Add [[nodiscard]] to all 35 enum_entries() ADL overloads

+++ test/source/test_config_file.cpp
@@ -156,7 +156,7 @@ TEST_CASE("load_gtopt_config - Reads [gtopt] section via GTOPT_CONFIG")
   REQUIRE(opts.threads.has_value());
   CHECK(*opts.threads == 8);
   REQUIRE(opts.algorithm.has_value());
-  CHECK(*opts.algorithm == static_cast<int>(LPAlgo::dual));
+  CHECK(*opts.algorithm == LPAlgo::dual);
   REQUIRE(opts.output_format.has_value());
   CHECK(*opts.output_format == "csv");
   REQUIRE(opts.output_compression.has_value());
